Allow running on a read-only filesystem

+++ docker/run.sh
@@ -4,6 +4,7 @@ set -e
 BASE_URL=${BASE_URL:-/}
 NGINX_ROOT=/usr/share/nginx/html
 INDEX_FILE=$NGINX_ROOT/index.html
+READONLY_FILESYSTEM=${READONLY_FILESYSTEM:-}
 
 node /usr/share/nginx/configurator $INDEX_FILE
 
@@ -23,13 +24,15 @@ replace_or_delete_in_index () {
   fi
 }
 
-if [ "${BASE_URL}" ]; then
+if [[ "${BASE_URL}" && -z $READONLY_FILESYSTEM ]]; then
   sed -i "s|location / {|location $BASE_URL {|g" /etc/nginx/nginx.conf
 fi
 
-replace_in_index myApiKeyXXXX123456789 $API_KEY
+if [[ -z $READONLY_FILESYSTEM ]]; then
+  replace_in_index myApiKeyXXXX123456789 $API_KEY
+fi
 
-if [[ -f $SWAGGER_JSON ]]; then
+if [[ -f $SWAGGER_JSON && -z $READONLY_FILESYSTEM ]]; then
   cp -s $SWAGGER_JSON $NGINX_ROOT
   REL_PATH="./$(basename $SWAGGER_JSON)"
   sed -i "s|https://petstore.swagger.io/v2/swagger.json|$REL_PATH|g" $INDEX_FILE
@@ -37,10 +40,12 @@ if [[ -f $SWAGGER_JSON ]]; then
 fi
 
 # replace the PORT that nginx listens on if PORT is supplied
-if [[ -n "${PORT}" ]]; then
+if [[ -n "${PORT}" && -z $READONLY_FILESYSTEM ]]; then
     sed -i "s|8080|${PORT}|g" /etc/nginx/nginx.conf
 fi
 
-find $NGINX_ROOT -type f -regex ".*\.\(html\|js\|css\)" -exec sh -c "gzip < {} > {}.gz" \;
+if [[ -z $READONLY_FILESYSTEM ]]; then
+  find $NGINX_ROOT -type f -regex ".*\.\(html\|js\|css\)" -exec sh -c "gzip < {} > {}.gz" \;
+fi
 
 exec nginx -g 'daemon off;'
